Fixes to -Wall compilation
By default, not all warnings are enabled. This change makes sure that
the project compiles correctly with following options turned on:

    -Wall -Wno-unused-but-set-variable -Wno-missing-braces \
    -Wno-unknown-pragmas -Wno-unused-function

+++ src/rct1/S4Importer.cpp
@@ -1508,7 +1508,7 @@ int S4Importer::GetSCNumber()
         utf8 digitBuffer[maxDigits + 1];
         utf8 * dst = digitBuffer;
         const utf8 * src = fileName + 2;
-        for (int i = 0; i < maxDigits && *src != '.'; i++)
+		for (size_t i = 0; i < maxDigits && *src != '.'; i++)
         {
             *dst++ = *src++;
         }
